3.5 Рефакторинг
