+++ src/utils.js
@@ -1,0 +1,34 @@
+// Функция генерации случайного целого числа из диапазона
+// Источник - https://github.com/you-dont-need/You-Dont-Need-Lodash-Underscore#_random
+export const getRandomInteger = (a = 1, b = 0) => {
+  const lower = Math.ceil(Math.min(a, b));
+  const upper = Math.floor(Math.max(a, b));
+
+  return Math.floor(lower + Math.random() * (upper - lower + 1));
+};
+
+export const getRandomArrayItems = (array) => {
+  const randomIndex = getRandomInteger(array.length - 1);
+
+  return array[randomIndex];
+};
+
+export const isExpired = (dueDate) => {
+  if (dueDate === null) {
+    return false;
+  }
+
+  let currentDate = new Date();
+  currentDate.setHours(23, 59, 59, 999);
+  currentDate = new Date(currentDate);
+
+  return currentDate.getTime() > dueDate.getTime();
+};
+
+export const isRepeatingTask = (repeating) => {
+  return Object.values(repeating).some(Boolean);
+};
+
+export const humanizeTaskDate = (date) => {
+  return date.toLocaleString(`en-GB`, {day: `numeric`, month: `long`});
+};
